Cleanup: Move each special_aftertrans_update to their respective TransData file
